+++ src/utils/utils.jl
@@ -71,7 +71,7 @@ function get_cep_variable_value(variable::OptVariable,
     for i in  1:length(index_set)
         if index_set[i]==Colon()
             push!(index_num,Colon())
-        elseif typeof(index_set[i])==Int64 || typeof(index_set[i])==UnitRange{Int64}
+        elseif typeof(index_set[i])==Int || typeof(index_set[i])==UnitRange{Int}
             push!(index_num,index_set[i])
         else
             new_index_num=findfirst(variable.axes[i].==index_set[i])
